add --check-bytecode to check lua bytecode after formatting.

+++ luastyle/core.py
@@ -3,9 +3,16 @@ import sys
 import time
 import subprocess
 import concurrent.futures
+from tempfile import mkstemp
 
 from luastyle.indenter import IndentRule, IndentOptions
 
+
+class BytecodeException(Exception):
+    def __init__(self, message):
+
+        # Call the base class constructor with the parameters it needs
+        super(BytecodeException, self).__init__(message)
 
 class Configuration:
     def load(self, filepath):
@@ -21,9 +28,10 @@ class Configuration:
 
 
 class FilesProcessor:
-    def __init__(self, rewrite, jobs, indent_options):
+    def __init__(self, rewrite, jobs, check_bytecode, indent_options):
         self._rewrite = rewrite
         self._jobs = jobs
+        self._check_bytecode = check_bytecode
         self._indent_options = indent_options
 
     def _process_one(self, filepath):
@@ -34,16 +42,22 @@ class FilesProcessor:
 
         rule_output = IndentRule(self._indent_options).apply(rule_input)
 
-        if self._rewrite:
-            f = open(filepath, 'r+')
-            f.seek(0)
-            f.write(rule_output)
-            f.truncate()
-            f.close()
+        if self._check_bytecode:
+            bytecode_equal = check_lua_bytecode(rule_input, rule_output)
         else:
-            print(rule_output)
+            bytecode_equal = True
 
-        return len(rule_output.split('\n'))
+        if bytecode_equal:
+            if self._rewrite:
+                f = open(filepath, 'r+')
+                f.seek(0)
+                f.write(rule_output)
+                f.truncate()
+                f.close()
+            else:
+                print(rule_output)
+
+        return bytecode_equal, len(rule_output.split('\n'))
 
     def run(self, files):
         print(str(len(files)) + ' file(s) to process')
@@ -62,7 +76,10 @@ class FilesProcessor:
             for future in concurrent.futures.as_completed(future_to_file):
                 file = future_to_file[future]
                 try:
-                    total_lines += future.result()
+                    success, n_lines = future.result()
+                    total_lines += n_lines
+                    if not success:
+                        raise BytecodeException('bytecode differs')
                 except Exception as exc:
                     print('%r generated an exception: %s' % (file, exc))
                 else:
@@ -75,17 +92,20 @@ class FilesProcessor:
 
 
 def check_lua_bytecode(raw, formatted):
-    currdir = os.path.dirname(__file__)
-
     if os.name == 'nt':
         raise NotImplementedError('check_lua_bytecode not supported on windows')
     else:
+        fd, raw_path = mkstemp()
+        fd, formatted_path = mkstemp()
+        fd, raw_bytecode = mkstemp()
+        fd, formatted_bytecode = mkstemp()
+
         # create raw file
-        with open(currdir + '/raw.lua', 'w') as file:
+        with open(raw_path, 'w') as file:
             file.write(raw)
 
         # create formatted file
-        with open(currdir + '/formatted.lua', 'w') as file:
+        with open(formatted_path, 'w') as file:
             file.write(formatted)
 
         # compile files, strip
@@ -94,12 +114,12 @@ def check_lua_bytecode(raw, formatted):
         else:
             luac = 'luac'
 
-        os.system(luac + ' -s -o ' + currdir + '/raw.out ' + currdir +'/raw.lua')
-        os.system(luac + ' -s -o ' + currdir + '/formatted.out ' + currdir + '/formatted.lua')
+        os.system(luac + ' -s -o ' + raw_bytecode + ' ' + raw_path)
+        os.system(luac + ' -s -o ' + formatted_bytecode + ' ' + formatted_path)
 
         # check diff
         # This command could have multiple commands separated by a new line \n
-        some_command = 'diff ' + currdir + '/raw.out ' + currdir + '/formatted.out'
+        some_command = 'diff ' + raw_bytecode + ' ' + formatted_bytecode
         p = subprocess.Popen(some_command, stdout=subprocess.PIPE, shell=True)
         (output, err) = p.communicate()
         # This makes the wait possible
@@ -108,12 +128,12 @@ def check_lua_bytecode(raw, formatted):
         output = output.decode("utf-8")
 
 
-        bytecode_equals = (p_status == 0) and (output == "")
+        bytecode_equal = (p_status == 0) and (output == "")
 
         # cleanup
-        os.remove(currdir + '/raw.lua')
-        os.remove(currdir + '/formatted.lua')
-        os.remove(currdir + '/raw.out')
-        os.remove(currdir + '/formatted.out')
+        os.remove(raw_path)
+        os.remove(formatted_path)
+        os.remove(raw_bytecode)
+        os.remove(formatted_bytecode)
 
-        return bytecode_equals
+        return bytecode_equal
